+++ core/silkworm/chain/genesis.hpp
@@ -25,7 +25,7 @@ namespace silkworm {
 * \brief Returns genesis data given a known chain_id.
 * If id is not recognized returns an invalid json string
 */
-std::string read_genesis_data(unsigned int chain_id);
+std::string read_genesis_data(unsigned long long chain_id);
 
 }  // namespace silkworm
